refactor: split env vars into client and server modules
- Separate environment variables into [src/utils/client/env.ts](cci:7://file:///Users/santoshvenkatraman/Personal/Coding/builder-prompt-ai/src/utils/client/env.ts:0:0-0:0) and [src/utils/server/env.ts](cci:7://file:///Users/santoshvenkatraman/Personal/Coding/builder-prompt-ai/src/utils/server/env.ts:0:0-0:0) to prevent leaky abstractions and runtime crashes
- Update [server/env.ts](cci:7://file:///Users/santoshvenkatraman/Personal/Coding/builder-prompt-ai/src/utils/server/env.ts:0:0-0:0) to use `process.env` for correct secrets loading in Node environment
- Update [client/env.ts](cci:7://file:///Users/santoshvenkatraman/Personal/Coding/builder-prompt-ai/src/utils/client/env.ts:0:0-0:0) to use `import.meta.env` for Vite compatibility
- Update `convex-client` and `router` to use `client/env` with `VITE_CONVEX_URL`
- Update `analyze-prompt` and `MixpanelProvider` to use `server/env`
- Resolve "Invalid environment variables" crashes by ensuring correct runtime usage
